+++ lib_unprompted/shared.py
@@ -13,7 +13,7 @@ import time
 class Unprompted:
 	def __init__(self, base_dir="."):
 		start_time = time.time()
-		self.VERSION = "9.6.0"
+		self.VERSION = "9.7.0"
 
 		self.log(f"Loading Unprompted v{self.VERSION} by Therefore Games", False, "SETUP")
 		self.log("Initializing Unprompted object...", False, "SETUP")