+++ shortcodes/stable_diffusion/remember.py
@@ -11,3 +11,6 @@ class Shortcode():
 
 	def cleanup(self):
 		self.globals = []
+
+	def ui(self, gr):
+		gr.Textbox(label="Arbitrary variable names to remember 🡢 verbatim", max_lines=1, placeholder="var_a var_b var_c")