+++ docs/CHANGELOG.md
@@ -3,7 +3,18 @@ All notable changes to this project will be documented in this file.
 
 For more details on new features, please check the [Manual](./MANUAL.md).
 
-<details open><summary>9.6.0 - 20 July 2023</summary>
+<details open><summary>9.7.0 - 20 July 2023</summary>
+
+### Added
+- New shortcode `[round]`: Rounds the first parg to a certain level of precision, e.g. 1.345 at `_place=1` yields 1.3, and 1644 at `_place=2` yields 1600
+- `[remember]`: Added Wizard UI
+- New WebUI Template Editor tab by o0oradaro0o (PR #146)
+- Minor UI updates for the Template Editor after merging
+- New config setting `stable_diffusion.template_editor`: you can set to `false` to disable the Template Editor tab
+
+</details>
+
+<details><summary>9.6.0 - 20 July 2023</summary>
 
 ### About
 This update resolves a number of issues related to `batch_index` evaluation, which were causing degraded image quality with `[zoom_enhance]`.